refactor: turn escrow account monitoring into an eventual
